Reorganize the test suite
